Move Balloon implementation to separate free-standing functions
